+++ include/IS_DC_Protocol.h
@@ -65,14 +65,15 @@ class AddStepInputDialog: public AddStepDialog {
 class ProtocolStep {
 	public:
 		enum stepType_t { PACE, SCALE, WAIT, STARTMODEL, STOPMODEL, RESETMODEL, CHANGEMODEL } stepType;    
-		int BCL; // ms
+		double BCL; // ms
+//		int BCL; // ms
 		int numBeats;
 		std::string currentToScale; // String name of current
 		int scalingPercentage; // Whole number %
 		int waitTime; // ms
 		enum modelType_t { LIVRUDY2009, FABERRUDY2000 } modelType;
 
-		ProtocolStep( stepType_t, int, int, std::string, int, int, modelType_t );
+		ProtocolStep( stepType_t, double, int, std::string, int, int, modelType_t );
 		~ProtocolStep( void );
 		int stepLength ( double );
 };
